@3.0.0 @Erased code related to HTTP server and refactored email sending logic to be more explicit with a dry-run mode. Updated tests and documentation accordingly.
