+++ index.ts
@@ -1,4 +1,4 @@
-import { trySendMail } from "./controllers/emailController";
+import sendMail, { EmailData } from "./controllers/emailController";
 import { getCodingFlavourEmail } from "./helpers/emailHelper";
 import SUBJECTS from "./helpers/subjectsHelper";
 import TEMPLATES from "./helpers/templatesHelper";
@@ -6,6 +6,7 @@ import SendGrid from "./services/emailService";
 import GmailService from "./services/gmailService";
 
 export {
-  getCodingFlavourEmail, GmailService, SendGrid, SUBJECTS, TEMPLATES, trySendMail
+  EmailData,
+  getCodingFlavourEmail, GmailService, SendGrid, sendMail, SUBJECTS, TEMPLATES
 };
