+++ jest.config.ts
@@ -6,11 +6,6 @@ export default {
   transform: {
     '^.+\\.ts$': ['ts-jest', { useESM: true }],
   },
-  globals: {
-    'ts-jest': {
-      useESM: true,
-    },
-  },
   moduleNameMapper: {
     '^(\.{1,2}/.*)\.js$': '$1',
   },
