+++ package.json
@@ -1,18 +1,14 @@
 {
   "name": "@coding-flavour/smtp",
-  "version": "2.0.0",
+  "version": "3.0.0",
   "main": "./index.ts",
-  "bin": "./src/app.ts",
   "type": "module",
   "scripts": {
-    "test": "npm run test:unit && npm run test:integration",
-    "test:unit": "jest test/unit --silent",
-    "test:integration": "jest test/integration --silent",
-    "dev": "nodemon ./src/app.ts",
-    "preprod": "ts-node ./src/app.ts --transpile-only",
     "companion:install": "tsx ./node_modules/@coding-flavour/companion/src/bin/companion.ts install --type=back",
-    "build": "tsc",
-    "postbuild": "tsc-esm-fix --target=dist --ext=\".js\""
+    "test": "npm run test:unit && npm run test:integration",
+    "test:integration": "jest test/integration --silent",
+    "test:unit": "jest test/unit --silent",
+    "test:unit:debug": "jest test/unit"
   },
   "keywords": [],
   "author": "",
@@ -20,24 +16,16 @@
   "description": "SMTP service for Coding Flavour",
   "dependencies": {
     "@coding-flavour/logger": "1.0.3",
-    "@sendgrid/mail": "^7.7.0",
-    "cors": "^2.8.5",
-    "dotenv": "^16.3.1",
-    "express": "^4.18.2",
-    "nodemailer": "^7.0.13"
+    "@sendgrid/mail": "7.7.0",
+    "nodemailer": "7.0.13"
   },
   "devDependencies": {
     "@coding-flavour/companion": "3.1.0",
-    "@types/cors": "^2.8.14",
-    "@types/express": "^4.17.17",
     "@types/jest": "^29.5.14",
-    "@types/nodemailer": "^7.0.9",
-    "jest": "^29.7.0",
-    "nodemon": "^3.0.1",
-    "ts-jest": "^29.3.2",
-    "ts-node": "^10.9.1",
-    "tsc-esm-fix": "^3.1.2",
-    "tsconfig-paths": "^4.2.0",
-    "typescript": "^5.2.2"
+    "@types/nodemailer": "7.0.9",
+    "jest": "29.7.0",
+    "ts-jest": "29.3.2",
+    "tsconfig-paths": "4.2.0",
+    "typescript": "5.2.2"
   }
 }
