+++ readme.md
@@ -2,7 +2,3 @@
 
 - Lee la documentación en [español](docs/README/README-ES.md)
 - Read the documentation in [english](docs/README/README-EN.md)
-
-# TODO
-
-- Añadir Coding Flavour Logger cuando tenga distribuible JS
