--- routes/emailRoutes.ts
@@ -1,8 +1,0 @@
-import { Express } from "express";
-import { sendMail } from "../controllers/emailController";
-
-const setRoutes = (app: Express) => {
-  app.post("/send-email", sendMail);
-};
-
-export { setRoutes };
