+++ types/Email.ts
@@ -1,0 +1,9 @@
+interface EmailData {
+  from: string;
+  to: string;
+  name: string;
+  message: string;
+  templateKey?: string;
+}
+
+export { EmailData };
